requirements.txt
bcrypt==3.2.2
cffi==1.15.1
cli-helpers==2.2.1
click==8.0.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
Flask==2.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pendulum==2.1.2
pgcli==3.4.1
pgspecial==1.13.1
Pillow==9.3.0
prompt-toolkit==3.0.30
psycopg2==2.9.1
pycparser==2.21
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2022.6
pytzdata==2020.1
rich==12.6.0
setproctitle==1.3.2
six==1.16.0
sqlparse==0.4.2
tabulate==0.8.10
wcwidth==0.2.5
Werkzeug==2.0.1
